extract customobject package reusable between solutions
