+++ customobject/postgresql_config.go
@@ -1,9 +1,18 @@
 package customobject
 
+import "fmt"
+
 // PostgreSQLConfig is custom object of postgresqlconfigs.containerconf.de custom
 // resource.
 type PostgreSQLConfig struct {
 	Spec PostgreSQLConfigSpec `json:"spec"`
+}
+
+func (m PostgreSQLConfig) Validate() error {
+	if err := m.Spec.Validate(); err != nil {
+		return fmt.Errorf("spec is not valid: %s", err)
+	}
+	return nil
 }
 
 // PostgreSQLConfigSpec is custom object specification. Represents the desired state
@@ -14,4 +23,14 @@ type PostgreSQLConfigSpec struct {
 	Database string `json:"database"`
 	// Owner is the database owner.
 	Owner string `json:"owner"`
+}
+
+func (s PostgreSQLConfigSpec) Validate() error {
+	if s.Database == "" {
+		return fmt.Errorf("database is not set")
+	}
+	if s.Owner == "" {
+		return fmt.Errorf("owner is not set")
+	}
+	return nil
 }
